services/arena/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
redis==5.0.1
azure-cognitiveservices-speech==1.35.0
boto3==1.34.40
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
